Refactor quake components.
