+++ src/js/features/Quake/ColoredFrags.jsx
@@ -1,19 +1,5 @@
 import React from "react";
-import { quakeTextToHtml } from "../../util/text";
 import { TextPulse } from "../Animations/Text.jsx";
-
-export const QuakeText = React.memo((props) => {
-  const { text, tag, color, className } = props;
-  const TagName = `${tag}`;
-  const textAsHtml = color ? quakeTextToHtml(text, color) : text;
-
-  return (
-    <TagName
-      {...{ className }}
-      dangerouslySetInnerHTML={{ __html: textAsHtml }}
-    />
-  );
-});
 
 export const ColoredFrags = React.memo((props) => {
   const { tag, frags, colors } = props;
