+++ src/js/features/Servers/Server.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import FavoriteToggle from "./FavoriteToggle.jsx";
 import { Scoreboard } from "./Scoreboard.jsx";
-import { QuakeText } from "../Quake/Quake.jsx";
+import { QuakeText } from "../Quake/QuakeText.jsx";
 import { copyToClipboard } from "../../util/clipboard.js";
 import { pluralize } from "../../util/text.js";
 import { TextBlur } from "../Animations/Text.jsx";
@@ -208,7 +208,8 @@ const ServerFooter = (props) => {
     <div className="server-footer p-3">
       <SpectatorButtons server={server} />
 
-      <div className="columns is-mobile is-vcentered app-text-small is-multiline">
+      <div
+        className="columns is-mobile is-vcentered app-text-small is-multiline">
         <div className="column">
           <div
             className="server-address"
